requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dj-database-url==2.2.0
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
packaging==24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
stripe==9.5.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.30.0
whitenoise==6.6.0
